+++ test/src/PageWidgetMock.cpp
@@ -14,7 +14,7 @@ void PageWidgetMock::emitPush(QVariant item)
     emit pushToStack(item);
 }
 
-void PageWidgetMock::emitRead(size_t index, QVariant &value)
+void PageWidgetMock::emitRead(int index, QVariant &value)
 {
     emit readFromStack(index, value);
 }
